# goios-peer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/device/{udid}/activate": {
            "post": {
                "description": "Returns and error if activation fails. Otherwise  {\"message\":\"Activation successful\"}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific",
                    "activation"
                ],
                "summary": "Activate the device by udid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/device/{udid}/conditions": {
            "get": {
                "description": "Get a list of the available conditions that can be applied on the device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Get a list of available device conditions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/instruments.ProfileType"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/disable-condition": {
            "post": {
                "description": "Disable the currently active condition on a device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Disable the currently active condition on a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/enable-condition": {
            "put": {
                "description": "Enable condition on a device by provided profileTypeID and profileID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Enable condition on a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Identifier of the profile type, eg. SlowNetworkCondition",
                        "name": "profileTypeID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Identifier of the sub-profile, eg. SlowNetwork100PctLoss",
                        "name": "profileID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/info": {
            "get": {
                "description": "Returns all lockdown values and additional instruments properties for development enabled devices.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Get lockdown info for a device by udid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device udid",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/device/{udid}/pair": {
            "post": {
                "description": "Pair a device with/without supervision",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Pair a device with/without supervision",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Set if device is supervised - true/false",
                        "name": "supervised",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Supervision *.p12 file",
                        "name": "p12file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Supervision password",
                        "name": "supervision_password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/profiles": {
            "get": {
                "description": "get the list of installed profiles from the ios device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "get the list of profiles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/resetlocation": {
            "post": {
                "description": "Reset the changed device location to the actual one",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Reset the changed device location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
        "/device/{udid}/screenshot": {
            "get": {
                "description": "Takes a png screenshot and returns it.",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Get screenshot for device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device udid",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    }
                }
            }
        },
        "/device/{udid}/setlocation": {
            "post": {
                "description": "Change the current device location to provided latitude and longtitude",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "general_device_specific"
                ],
                "summary": "Change the current device location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location latitude",
                        "name": "latitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Location longtitude",
                        "name": "longtitude",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Device UDID",
                        "name": "udid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.GenericResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.GenericResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "instruments.Profile": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "instruments.ProfileType": {
            "type": "object",
            "properties": {
                "activeProfile": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isDestructive": {
                    "type": "boolean"
                },
                "isInternal": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "profiles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/instruments.Profile"
                    }
                },
                "profilesSorted": {
                    "type": "boolean"
                }
            }
        }
    }
}